:bug: 新增团队资源支持并优化角色权限管理
- 新增团队资源管理接口和实现
- 优化资源查询逻辑，支持主库和团队库切换
- 重构角色权限管理，使用casbin进行权限控制
- 优化数据库连接和缓存机制
- 调整API路径和命名规范

+++ cmd/server/palace/internal/biz/team_role.go
@@ -45,6 +45,18 @@ func (b *TeamRoleBiz) UpdateTeamRole(ctx context.Context, teamRole *bo.UpdateTea
 
 // DeleteTeamRole 删除团队角色
 func (b *TeamRoleBiz) DeleteTeamRole(ctx context.Context, id uint32) error {
+	// 查询角色， 如果存在关联关系， 不允许删除
+	teamRoleDo, err := b.teamRoleRepo.GetTeamRole(ctx, id)
+	if !types.IsNil(err) {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return merr.ErrorI18nToastRoleNotFound(ctx)
+		}
+		return merr.ErrorI18nNotificationSystemError(ctx).WithCause(err)
+	}
+	if len(teamRoleDo.Apis)+len(teamRoleDo.Members) > 0 {
+		return merr.ErrorI18nToastRoleHasRelation(ctx)
+	}
+
 	if err := b.teamRoleRepo.DeleteTeamRole(ctx, id); !types.IsNil(err) {
 		return merr.ErrorI18nNotificationSystemError(ctx).WithCause(err)
 	}
